scripts/selenium/requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
install==1.3.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pydub==0.25.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.8.3
selenium-recaptcha-solver==1.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SpeechRecognition==3.8.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
webdriver-manager==4.0.1
wsproto==1.2.0
